Reverting test changes
