--- testsuite/test_006_isol_C2H4_4proc_PBE0ERI/Conquest_coord
@@ -1,10 +1,0 @@
-22.676714   0.000000    0.000000    
-0.000000    22.676714   0.000000    
-0.000000    0.000000    22.676714   
-6
-    0.50000000    0.50000000    0.44437500   2 T T T
-    0.50000000    0.50000000    0.55562500   2 T T T
-    0.57859183    0.50000000    0.39900000   1 T T T
-    0.57859183    0.50000000    0.60100000   1 T T T
-    0.42140817    0.50000000    0.60100000   1 T T T
-    0.42140817    0.50000000    0.39900000   1 T T T
