--- testsuite/test_007_isol_C2H4_4proc_PBE0ERI_GTO/C.gto
@@ -1,14 +1,0 @@
-C
-3
-2S  3   2.0   0
-      3.0500154803     -0.8718503385      0.0000000000
-      0.1599863956      0.3605314379      0.0000000000
-      0.5242567425      0.8731953379      0.0000000000
-2P  3   2.0   0
-      0.1720417090      0.1855258456      0.0000000000
-      2.7272258524      1.7315939925      0.0000000000
-      0.7055493743      0.8775500569      0.0000000000
-3D  3   0.0   1
-     86.0321546336      0.6114404190      0.0000000000
-      3.6206685167      1.5850293864      0.0000000000
-      0.5229417354      0.5889629298      0.0000000000
