--- testsuite/test_007_isol_C2H4_4proc_PBE0ERI_GTO/H.gto
@@ -1,6 +1,0 @@
-H
-1
-1S  3   1.0   0
-      0.5275548833      0.6715259742      0.0000000000
-      2.7696756637      0.5826356048      0.0000000000
-      0.1171533321      0.3049461604      0.0000000000
